Update GitHub build actions and switch to Yarn (#3)
* Add GitHub build actions and update existing release workflow.
* Add '.gitattributes' to allow more deterministic builds across platforms.
* Fix various documentation issues including logo in readme, 'markdownlint' warnings, and minor formatting fixes.
* Fix and update workflows, scripts, and package management, including allowing different package managers in Makefile, enabling `corepack`, and setting executable bit on tool scripts.
* Refactor scripts for better readability and consistency.
* Add pre-commit config that checks YAML syntax and runs 'checkmake' to check 'Makefile'.
* Update `VSCode` and `ESLint` settings, reformat various files, and fix warnings in Python and shell scripts.

+++ tools/import-crowdin.sh
@@ -2,39 +2,40 @@
 #
 # This script assumes a linux environment
 
-set -e
+set -eEax
 
 echo "*** uBlock: Importing from Crowdin archive"
 
 SRC=~/Downloads/crowdin
-rm -r $SRC || true > /dev/null
-unzip -q ~/Downloads/uBlock\ \(translations\).zip -d $SRC
+[[ -d "$SRC" ]] && rm -r $SRC
+ARCHIVE=~/Downloads/uBlock\ \(translations\).zip
+[[ -e "$ARCHIVE" ]] && unzip -q "$ARCHIVE" -d $SRC
 
 # https://www.assertnotmagic.com/2018/06/20/bash-brackets-quick-reference/
 
 DES=./src/_locales
 DESMV3=./platform/mv3/extension/_locales
 
-for dir in $SRC/*/; do
-  srclang=$(basename $dir)
-  deslang=${srclang/-/_}
-  deslang=${deslang%_AM}
-  deslang=${deslang%_ES}
-  deslang=${deslang%_IN}
-  deslang=${deslang%_LK}
-  deslang=${deslang%_NL}
-  deslang=${deslang%_PK}
-  deslang=${deslang%_SE}
-  if [[ $deslang == 'en' ]]; then
-    continue
-  fi
-  # ubo
-  mkdir -p "$DES/$deslang/" && cp "$SRC/$srclang/messages.json" "$DES/$deslang/"
-  # ubo lite
-  mkdir -p "$DESMV3/$deslang/" && cp "$SRC/$srclang/uBO-Lite/messages.json" "$DESMV3/$deslang/"
-  # descriptions
-  #cp "$SRC/$srclang/description.txt" "./dist/description/description-${deslang}.txt"
-  cp "$SRC/$srclang/uBO-Lite/webstore.txt" "./platform/mv3/description/webstore.$deslang.txt"
+for dir in "$SRC"/*/; do
+    srclang=$(basename "$dir")
+    deslang=${srclang/-/_}
+    deslang=${deslang%_AM}
+    deslang=${deslang%_ES}
+    deslang=${deslang%_IN}
+    deslang=${deslang%_LK}
+    deslang=${deslang%_NL}
+    deslang=${deslang%_PK}
+    deslang=${deslang%_SE}
+    if [[ $deslang == 'en' ]]; then
+        continue
+    fi
+    # ubo
+    [[ -e "$SRC/$srclang/messages.json" ]] && mkdir -p "$DES/$deslang/" && cp "$SRC/$srclang/messages.json" "$DES/$deslang/"
+    # ubo lite
+    [[ -e "$SRC/$srclang/uBO-Lite/messages.json" ]] && mkdir -p "$DESMV3/$deslang/" && cp "$SRC/$srclang/uBO-Lite/messages.json" "$DESMV3/$deslang/"
+    # descriptions
+    #cp "$SRC/$srclang/description.txt" "./dist/description/description-${deslang}.txt"
+    [[ -e "$SRC/$srclang/uBO-Lite/webstore.txt" ]] && cp "$SRC/$srclang/uBO-Lite/webstore.txt" "./platform/mv3/description/webstore.$deslang.txt"
 done
 
 # Output files with possible misuse of `$`, as this can lead to severe
@@ -45,7 +46,6 @@ echo "*** uBlock: Instances of '\$':"
 grep -FR "$" $DES/ || true
 grep -FR "$" $DESMV3/ || true
 
-
-rm -r $SRC
+[[ -d "$SRC" ]] && rm -r $SRC
 echo "*** uBlock: Import done."
 git status
